+++ packages/moengage_flutter/moengage_flutter_platform_interface/lib/src/utils/utils.dart
@@ -171,3 +171,8 @@ Map<String, dynamic> filterMapWithSupportedTypes(Map<String, dynamic> data) {
   });
   return filteredMap;
 }
+
+/// Return whether type of [identity] is supported or not
+bool isSupportedIdentity(dynamic identity) {
+  return identity is String || identity is Map<String, dynamic>;
+}