+++ packages/moengage_flutter/moengage_flutter_ios/CHANGELOG.md
@@ -1,5 +1,10 @@
 # MoEngage Flutter iOS Plugin
 
+# Release Date
+
+## Release Version
+- Adding support for identify user based on multiple identities
+
 # 18-02-2025
 
 ## 3.2.1